Updated code samples, more docs
